requirements.txt
carla==0.9.15
hydra-core==1.3.2
keyboard==0.13.5
matplotlib==3.4.0
mmdet==3.2.0
numpy==1.23.3
omegaconf==2.3.0
opencv_python==4.6.0.66
Pillow==10.1.0
pygame==2.5.2
pyquaternion==0.9.9
tensorboardX==2.6.2.2
timm==0.9.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
